+++ domain/src/main/java/com/remindme/domain/usecase/alarm/implementation/CancelAlarmImpl.kt
@@ -5,13 +5,13 @@ import com.remindme.domain.repository.TaskRepository
 import com.remindme.domain.usecase.alarm.CancelAlarm
 import javax.inject.Inject
 
-class CancelAlarmImpl @Inject constructor(
+ class CancelAlarmImpl @Inject constructor(
     private val taskRepository: TaskRepository,
     private val alarmInteractor: AlarmInteractor
 ) : CancelAlarm {
 
 
-    override suspend fun invoke(taskId: Int?) {
+    override suspend fun invoke(taskId: Long?) {
         val task = taskRepository.findTaskById(taskId) ?: return
 
         val updatedTask = task.copy(dueDate = null)
