+++ features/glance/src/main/java/com/remindme/glance/model/Task.kt
@@ -2,4 +2,4 @@ package com.remindme.glance.model
 
 import javax.inject.Inject
 
-data class Task @Inject constructor(val id: Int?, val title: String)
+data class Task @Inject constructor(val id: Long?, val title: String)
